# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This Api is used for creating and managing giftcards",
        "title": "Gift Card API",
        "termsOfService": "NONE",
        "contact": {
            "name": "info@jabama.com",
            "url": "jabama.com",
            "email": "info@jabama.com"
        },
        "license": {
            "name": "NONE",
            "url": "None"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/campaign": {
            "put": {
                "description": "updates a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "updates a campaign",
                "operationId": "update",
                "parameters": [
                    {
                        "description": "Update Campaign dto",
                        "name": "campaignDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCampaignDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            },
            "post": {
                "description": "store a new campaign and generates the keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "store a campaign",
                "operationId": "store",
                "parameters": [
                    {
                        "description": "Create campaign dto",
                        "name": "campaignDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCampaignDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/campaign/page/{size}/{number}": {
            "get": {
                "description": "get list of campaigns in paging object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign paging",
                "operationId": "find-page",
                "parameters": [
                    {
                        "type": "number",
                        "description": "page size",
                        "name": "size",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "page number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "search by title",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignPageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/campaign/{id}": {
            "delete": {
                "description": "deletes a campaign by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "deletes a campaign",
                "operationId": "delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "campaign's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteMessageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
        "/v1/gift-card/page/{size}/{number}": {
            "get": {
                "description": "get list of gift cards in paging object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "gift cards paging",
                "operationId": "find-page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "size",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "campaign id",
                        "name": "campaignId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search in public key",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is valid gift card",
                        "name": "isValid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "expire date from",
                        "name": "expireDateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "expire date to",
                        "name": "expireDateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardsPageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CampaignDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.CampaignPageDTO": {
            "type": "object",
            "properties": {
                "campaigns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CampaignDTO"
                    }
                },
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateCampaignDTO": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeleteMessageDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GiftCardDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "campaign_id": {
                    "type": "integer"
                },
                "campaign_title": {
                    "type": "string"
                },
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "is_valid": {
                    "type": "boolean"
                },
                "public_code": {
                    "type": "string"
                },
                "secret_code": {
                    "type": "string"
                },
                "uun": {
                    "type": "string"
                }
            }
        },
...
        "dto.GiftCardsPageDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "gift_cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GiftCardDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
        "dto.UpdateCampaignDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "indraframework.IndraException": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "severity": {
                    "type": "integer"
                },
                "technicalMessage": {
                    "type": "string"
                }
            }
        }
    },
...
}